#414: more refactoring
