+++ tendermint/tests/lite.rs
@@ -7,7 +7,7 @@ use tendermint::lite::{Requester, TrustThresholdFraction, TrustedState};
 use tendermint::{block::signed_header::SignedHeader, evidence::Duration, lite, Hash, Time};
 
 mod utils;
-use utils::lite::*;
+use utils::{*, lite::*};
 
 #[derive(Deserialize, Clone, Debug)]
 struct TestCases {
